[perfetto_consumer] Improve Streaming Behaviour
When streaming is enabled, the trace-engine won't do a buffer swap while
the trace context is held. We need to instead hold the prolonged context
to ensure the trace session doesn't end, and repeatedly obtain and
release the trace context to allow for buffer swaps between writes.

This also adds a rust compatible context.copy_record function which
calls into trace_context_alloc_record. This allows us to check if the
buffer actually has room to write the record. If it doesn't, we
temporarily back off, ensuring we don't drop any of the perfetto data.

Bug: 421224263

Change-Id: Ie63a9d25ee000d9c68bbb1e3c0b50b2f6a0c50a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288786
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Jia Li <lizhijia@google.com>