[flatland] Refactor how args are passed to flatland::Renderer
There are several optional args that may be passed to flatland::Renderer
in no particular order.  Introduces a new struct Renderer::RenderArgs
that is passes as a non-optional argument to Renderer::Render().

This is in preparation for two additional features, each of which adds
an additional optional arg.

Bug: 423066970
Bug: 423071751
Change-Id: I815d1abdc387cdb407bc15272319946e2bdbb569
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295167
Reviewed-by: Justin Kobler <jkobler@google.com>
Fuchsia-Auto-Submit: Josh Gargus <jjosh@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>