[netstack] Downgrade Interfaces Admin Error Log
The go FIDL bindings have a race that occasionally leads to failing to
send address removal reasons. Because the bindings are only used by
Netstack2 and Netstack2 is in in maintenance mode, we're not investing
in fixing the underlying issue. Instead, downgrade the log level when
this event happens, as logging "error" leads to flaking integration
tests.

Bug: 423674861
Change-Id: If99cb3b9696372ce770d0b21d825e8831b6433a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1296644
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>