[component_manager] aggregate exposes by name and target, not just name
When we aggregate together exposes we need to consider both the name and
target for this aggregation step, as there are now two valid expose
targets (parent and framework). Without this if a component exposes two
capabilities with the same target name to different sources we run the
risk of aggregating together two capabilities that should not be grouped
like that, and the violation of this assumption results in a crash.

Fixed: 422528939
Change-Id: Ia0510150e08ce1f7c9b88c300765c0341003c2a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1294625
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>
Fuchsia-Auto-Submit: Claire Gonyeo <cgonyeo@google.com>