[kernel][arm64] Formalize arm64 ZBI protocol cache guarantees (reland)
This is a reland of fxrev.dev/1266427. The change employs some trickery
to associate extern C++ symbols with labels in an inline assembly block.
Initially the latter labels were local, and that was found to present
issues under LTO (hence the revert), where they would compile away and
not be present at C++ link time. The diff with this patch was to keep
the label logic within the function and away from link-time
considerations.

Spurred by the desire to remove set/way data cache operation usage, we
turn to the current, primary, de facto use-case for these: in our
present implementation the arm64 ZBI booting protocol, we use these as a
part of a routine to flush the whole boot CPU data cache before
disabling it (sparing a naive ZBI kernel any surprises when it
re-enables it). This behaviour wasn't spec'd, but seemed like a safe
and reasonable thing to do for the time being. Without set/way
operations at our disposal, we no longer have an equivalent to that
routine. So we weaken and finally spec the guarantees.

For pragmatic reasons discussed elsewhere, we ultimately adopt what
Linux requires of its bootloaders:
"""
The address range corresponding to the loaded kernel image must be
cleaned - but not necessarily invalidated - to the point of coherency
in the data cache (reserve memory included) and invalidated in the
instruction cache (reserve memory not necessarily included).
""""

With no data range guaranteed as clean beyond the kernel's load image,
ZBI kernels must now be extra careful with access to the data ZBI and
physical memory allocations before it re-enables the data cache. We
leverage already-prepared carveouts around EarlyBootZbi and Pool access
callbacks for this.

In the case of trampoline booting booting a kernel and ZBI that overlap
with the current load image, it is difficult to guarantee that we do not
accidentally generate dirty cache lines corresponding that next loaded
images. That is, unless we do a full (set/way) local cache flush. As a
switchback to removing this type of flushing elsewhere, we introduce a
trampoline-boot-specific override to BootZbi::ZbiBoot() that still
employs it. Eventually we'll remove the ability to load images that
overlap with the current load image (which is a feature only exercised
in tests) and remove this override.

Bug: 408020980
Change-Id: I308746946b0d87989fad10219fed3a0f5bbbdcdb
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1292945
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>