[assembly] Don't copy board_script_path twice
For zbi signing, boards can provide a signing script, and we require
that it's also listed in the script inputs. This leaks the same path
twice into the board config, which causes bad behavior when
constructing the board config.

The specific bad behavior is that when we use hard-links, the file
somehow gets emptied when writing the board config to a directory.

Change-Id: I309d0c523d03ca3f80d3234b9732369fe89dbf23
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1292709
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>