[starnix] Add mmap example with a sample record
Add an example PERF_RECORD_SAMPLE. Usually after the metadata, there are a bunch of samples. The type of sample depends on the header. In this case, we get the header for the type PERF_RECORD_SAMPLE (either for `misc` field of USER or KERNEL).

Bug: 398914921

Change-Id: I6be2fa21500ba648c960487c63b427cf79d517f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276924
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Mary Xia <maryxia@google.com>