[df][irq] Fix an issue with default IRQ modes.
It looks like driver framework code is in the process of transitioning
from an old mode, where drivers were responsible for providing
interrupt configuration, to a new mode where the configuration should
be known by the platform driver.

During the transition, it looks like code is expected to pass a value
of 0 for the flags, to indicate that the platform configuration should
be used instead.  Otherwise, the user's configuration for the
interrupt mode will be used.

Except, that is not entirely correct either.  There was code in here
which was also trying to see if platform config had specified that an
IRQ should be a wake vectors as well, and was trying to enable the
wake vector flag in the flags eventually sent to
`zx_interrupt_create`, even if the user supplied value was non-zero
and we were meant to be using the user-config.

Sadly, this code was incorrect as well, as it was written as:

```
  if (bus_->suspend_enabled() &&
      irq.wake_vector().has_value() &&
      irq.wake_vector().value()) {
    flags &= ZX_INTERRUPT_WAKE_VECTOR;
  }
```

Note that `flags &= ZX_INTERRUPT_WAKE_VECTOR` is not going to be
setting the flag, instead it is going to clear *all* of the flags
except the WAKE_VECTOR flag (if it was set to begin with).  This was
supposed to be `|=` instead.  The only saving grace here seems to be
that this logic was never actually triggering; I guess no IRQs so far
have been configured as wake vectors at the platform bus IRQ level?

All of this was leading to a problem where the USB driver was trying
to fetch IRQ 287, and passing the value `ZX_INTERRUPT_WAKE_VECTOR` as
their only config.  The level and mode were being left as 0x0 (aka,
"default").  Code in the platform driver saw a non zero value passed
for the flags and just used the config passed by the USB driver.  This
went to the kernel as default, which means that the GIC driver chose
the mode, which turned out to be hardcoded to Edge/Active-High.  This
is incorrect for the USB driver; this is a level triggered interrupt,
not an edge triggered interrupt.

So, make an attempt to clean up at least some of this while the
transition is happening.
+ If the user passes a 0 value for the interrupt *mode*, then replace
  the interrupt *mode* with what we have in config.  IOW - don't
  demand that _no_ flags were set in order to user the config in the
  platform driver, just the bits which describe the mode.
+ If the user passes a non-zero value for the mode config, print a
  warning if that disagrees with what we have in the platform config,
  but continue to use the user's config.
+ Fix the WAKE_VECTOR flagging statement to actually set the flag,
  instead of clearing all of the other flags.

Bug: 423687236
Change-Id: Ie676d3966baf804ff2aa1349a843881f7761d07f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1294323
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: John Grossman <johngro@google.com>