[early-boot-instrumentation] ZBI_ITEM_TYPE_DEBUGDATA support
This just adds function and unit-tests. The integration will follow on a
separate patch. The approach consider the contents to be mandatory,
while the logs may be empty.

For logs, when present, they are added to the log output dir, where the
name is a concatenation of the sink_name-vmo_name.log.

Any error to set a property the vmo(we reuse the returned VMO, such that
we only remove the trailing parts of the vmo), is mentioned in the logs
and the vmo is still exposed.

Bug: 42071544
Change-Id: I4cf9342a25fd741ab85b698c6bc76a137a3a8473
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283886
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>