[msd-arm-mali] Cancel context after two terminated atoms
The ICD is supposed to respond to a terminated atom by
closing the context. If that does not happen then
close the context on our end.

We have been seeing infrequent reports of devices that are
logging terminated atoms repeatedly. This change will close
their connections earlier and hopefully provide a report
closer to the time that the hardware gets into a bad state.

Bug: 414378962
Change-Id: Id6784da12f24c43431948d27516dcc26343eaa80
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1291784
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>