[kernel][phys][vm] Re-formulize temporary hand-off pointers
Previously, pointers in the hand-off were accomplished by a thin wrapper
around physical addresses that passed those into the physmap on
'dereference'. This was meant as a interim measure until we'd be in a
position to just pass plain, raw pointers. The missing link was the
ability to create mappings in physboot and plumb their descriptions and
'registration' to the kernel VM. We now have those things.

We re-formulize temporary hand-off pointers accordingly, keeping the
wrapper class for type safety (especially distinguishing them from the
soon-to-be-introduced permanent hand-off pointer type) as well as a
place to assert that gPhysHandoff is unset, ensuring that these pointers
are not dereferenced after the end of hand-off.

The mappings that back temporary hand-off memory are conveyed to the VM
and end up in a single 'temporary' VMAR that gets destroyed at the end
of hand-off.

We introduce here as well an allocator of virtual address ranges for use
in hand-off mappings. The address space population strategy here is as
follows (taking the kernel's load address as an input):

* The physmap is a fixed mapping below the rest of the others

* Virtual addresses for temporary and permanent hand-off data are
  bump-allocated downward (in 1GiB-separated ranges) below the kernel's
  memory image (wherever it was loaded)

* The remaining, various first-class mappings are made just above the
  kernel's memory image (wherever it was loaded), with virtual address
  ranges bump-allocated upward.

           ...
    other first-class     (↑)
         mappings
-------------------------
      hole (1 page)
-------------------------
   kernel memory image
-------------------------
      hole (1 page)
-------------------------
 permanent hand-off data  (↓)
------------------------- kernel load address - 1GiB
 temporary hand-off data  (↓)
           ...
------------------------- kArchPhysmapVirtualBase
         physmap
------------------------- kArchPhysmapVirtualBase + kArchPhysmapSize

We can always revisit this strategy later (e.g., if it proves
insufficient for, say, the degree of randomness we want in a KASLR
implementation), but it's good enough for now and its implementation
lives behind an API that shouldn't need updating if this were to change.

Bug: 42164859
Change-Id: I998b1985a7c54c8622f151a3a04dd5ce1ace6571
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224751
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>