[kernel][phys][vm] Formalize permanent hand-off pointers
Extending the machinery and formalism we have around handing off
pointers to temporary hand-off data backed by temporary mappings, we can
now do so similarly with permanent pointers to permananent data backed
by permanent mappings.

As established with the temporary machinery the allocation of permanent
virtual addresses happens via bump allocation effectively off the end of
the kernel's load image:

           ...
    other first-class     (↑)
         mappings
-------------------------
      hole (1 page)
-------------------------
   kernel memory image
-------------------------
      hole (1 page)
-------------------------
 permanent hand-off data  (↓)
------------------------- kernel load address - 1GiB
 temporary hand-off data  (↓)
           ...
------------------------- kArchPhysmapVirtualBase
         physmap
------------------------- kArchPhysmapVirtualBase + kArchPhysmapSize

Permanent hand-off mappings are packaged in a hand-off VMAR, which is
included in the list of VMARs passed to VM init.

As a demonstration of this new functionality, we make good on
long-standing TODOs to recast the boot options struct, version string,
and mexec data ZBI as permenanent hand-off data.

Bug: 42164859
Change-Id: Ifc3f9e8d0b557446443725de5cd6f0135691d5fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1224753
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>