[starnix] Inline FutexTable::load_futex_value.
This hides the atomic ordering actually in use for operations. As an
alternative we could add an ordering argument but then this method's
only purpose would be to provide an error type if called without a
memory manager. That error conversion is fairly concise, so might as
well just skip an additional bit of API surface.

Future changes may modify the atomic orderings here, but this is purely
mechanical.

Change-Id: Ie1bc9ee581b375a9cfc40fe82f6a3c3df3731e42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1297374
Commit-Queue: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>