[fxfs] Add new protocol for streaming blob volume
Add a new protocol for streaming a new blob volume to disk. The
implementation is still under development and does not modify any
existing directory entries. Currently it takes a payload from a VMO
representing an fxfs partition with a single blob volume, writes it to a
file in the existing blob volume, and mounts it as a nested filesystem
within the new volume streaming service.

A lot of things are currently hard-coded to match the test cases and
test fixtures, such as the volume name and the blobs we expect to find.
The protocol is currently tied to a blob volume when it is mounted, so
we may need to move things around so they work at the filesystem level.

Bug: 397515768
Test: fxfs-platform-test
Change-Id: Ic252ed18d7e9f4819403be5bc67b887e31e4bbc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1276944
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>