[component] Save memory by generating default Requests on demand
For most routers, component manager provisions it with a default request
object. This object was kept resident in memory separately for each
Router, which adds up to a large amount of memory. Instead, we can
compute the default request on demand if it's needed, because its
contents are knowable a priori.

This saves approx 1.25 MiB on core.

Bug: 419024317
Change-Id: I24cfe696377515e256fd41052be362f4ae8c6f32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1297464
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>