[bazel] Build json_validator_valico with Bazel.
This is an simple reland of [1], which includes a fix for the
issue that required a revert [2]. See the difference between patchset 1
and patchset 2 on the current Gerrit issue to see what it was.

This CL moves the json_validator_valico host tool entirely
to the Bazel graph, and makes it usable from GN actions using
the bazel_tool_action() template.

[1]: https://fuchsia-review.googlesource.com/c/fuchsia/1283064
[2]: https://fuchsia-review.googlesource.com/c/fuchsia/1296904

Bug: 418925543
Change-Id: I2de87245f63ebbac228d51af36b358fac069c616
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1298064
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>