[fxblob] Use make-blob-image for creating test images
Refactor make-blob-image so that it can be reused to create Fxblob
images using both an assembly manifest as well as in-memory blobs. This
allows us to reuse the implementation for testing in our Fxblob platform
tests, which previously would install delivery blobs using a pattern
more similar to OTA.

This helps promote more code reuse, and ensures that we are testing
images in the format we will create them in. This also avoids the
workaround previously added to avoid running fsck on the TestFixture
when it was backed using a ramdisk, due to the inability to reopen a
BlockDevice.

Bug: 397515768
Test: fx test fxfs-platform-test
Change-Id: I42bd2886883feb41bedb8b2572f12ba55eaa6c57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1298145
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>