[wlan][policy] Set country code after phy reset
A phy reset will, in the case of brcmfmac, reset the phy's country code.
Typically, we set the country code on a phy when it appears. However,
in brcmfmac, the phy will not go away during the reset process.
Therefore, we need to explicitly set the country code as part of reset.

Bug: b/424173437
Test: fx test wlancfg-tests; Antlion wlan_reboot_ap_open_test.sh
Change-Id: I56d82852208d69f3d531ac2df3522b519cde5cc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1299585
Fuchsia-Auto-Submit: Marc Khouri <mnck@google.com>
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Marc Khouri <mnck@google.com>