[assembly] Add "cipd_name" to board and bib_set metadatas.
In order to ensure all assembly artifacts have unique resource
names (in MOS), the assembly target names will be prefixed with
the type of artifact. E.g. board config targets will be named
"board.<target_name>".

In order to not affect existing customers, the existing CIPD
location for these artifacts will remain the same.

This CL introduces a "cipd_name" field to the Build APIs for
Board Configs and BIB sets, which is set to the current value that
the "name" field is set to, and infra will begin to use this to upload
to CIPD.

Bug: b/418249336
Change-Id: I04983df09a7f6e595fcdcf55c169c060600416f5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295088
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>