[minfs] bcache dies on failed mutation.
Bcache is the natural choke point of all disk operations in minfs, so
watching there to crash the filesystem if a mutation fails. There is an
option in minfs to disable this since it is not desirable for unit
tests.

Rearranges an integration test to be more tolerable of minfs going away
when the device stops accepting transactions.

Bug: 416027941, 416305709

Change-Id: I7f558ac7656420077fb211170cfbb3cffc222a19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1292805
Reviewed-by: Rui Qi Sim <rqsim@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>