[kernel][vm] Panic if PMM allocations wait too long
If PMM allocations are blocked for long, it's possible that the memory
watchdog or/and the evictor are wedged. Panic instead of remaining in an
unrecoverable state.

Add some diagnostic logs in the evictor that should hopefully start
printing before we panic.

Bug: 423900604
Run-All-Tests: true
Change-Id: I9be13b4f685aad6b8a48bdad6b7525052a491924
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1296925
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>