[timekeeper] Wait for hrtimer to enumerate
Rather than assume it will already be enumerated by the time timekeeper
starts, it is more correct to wait for it to enuemrate. In order to do
this safely, we need to know for sure whether the product/board expects
a device to even enumerate. Therefore we only attempt to connect if we
have been configured to think it should enumerate. This config is
currently specified via the product, but it should be specified via a
board config in the future.

Change-Id: Ia19cec64ea51319ff38cdfd70cc745e1a8e3e153
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1220286
Reviewed-by: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>