[ffx] Implement fho::SubtoolSuite
Create a mechanism for implementing ffx top-level subtools that are
compiled separately to dispatch to their subcommands. Consistent with
the other subtools, both the dispatcher and the dispatched subcommands
implement `FfxTool`. This mechanism does not use macros or jinja -- it
simply requires the subtool to implement a few lines of boilerplate,
and one function that creates the subtools depending on the subcommand

Change-Id: I6a6a69640725ed7f3d208fc6ffaa2a7c9f64bc86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1300145
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>