[aml-usb-phy] Change AddChild asserts to logs
The driver framework calls to AddChild, AddService etc are not expected
to fail in normal operations, but when the system shutdown is issued
while these calls were in progress, it can return errors. aml-usb-phy
was relying on the expectation that these calls will never fail and
there were asserts added for the same. It was noticed that sometimes the
system could be restarted before the driver start up is completed and
it resulted in crashes due to these asserts.
The assumption that these calls will never fail is incorrect.
We need to handle the error properly. Changed the asserts to error codes
which would result in driver unbind upon error.

Fixed: 416590086
Change-Id: Iecb8ae6b8e80220c320b68ea1414a8002cd38b44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1299885
Commit-Queue: Puneetha Ramachandra <puneetha@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>