[profiler] Allow events aggregation in profiler symbolize
Currently, we put all of the events belong to the same tid into a single
array. The issue of this method is that pprof itself will treat it as an
very long call stack, lack of the ability to aggregate them. We have to
create brand new sample for every call stack to allow aggregation in
pprof.This change modify the symbolized text to facilicate creating
brand new samples for every call stack.

Bug: 401034854

Change-Id: Ibb42a0ee3c1636eebe68eb1361dce96989b610d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1293652
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Jia Li <lizhijia@google.com>