[kernel][vm] Implement VmPageSpliceList with VmPageList
Instead of having its own WAVL tree, skew and start + end nodes this
change uses the existing VmPageList data structure instead. A side
effect of this change is that TakePages can now fail due to a memory
allocation failure, as the start and end nodes are not preallocated.
This failure is tolerable, however, as the usage locations are already
tolerant to failures.

The benefit of using a VmPageList is that existing VmPageList methods
can be used to manipulate the splice list, instead of having to re-roll
them again against a raw WAVL tree.

This benefit is partially utilized in this commit to simplify the
implementation of the existing VmPageSpliceList methods, however it
will more importantly allow trivially adding additional functionality
to the VmPageSpliceList, as the VmPageList already supports a rich set
of operations.

Change-Id: Ide129cb132037a8877a15e5a66e6cb2f2b04bdc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1299006
Commit-Queue: Adrian Danis <adanis@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>