[build][idk] Only validate schema when generating IDK
Remove validation of the schema for IDK meta.json manifests from
sdk_atom() and sdk_collection(). The schema of manifests in the IDK is
now validated when generating the IDK. The schema of manifests for atoms
not in the IDK, such as FIDL libraries in the "prebuilt" category are no
longer validated, but this is okay because we do not publish them.

Validation of the IDK build manifests remains unchanged.

Bug: 413071161
Change-Id: I3354bd5d9adf94c0c7334e063aa80421113e02fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1287649
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>