[ffx][audio] Fix 'ffx audio play' position
Rewrite the low-level loop for `ffx audio play`, for both
correctness and clarity. Eliminate float-based mechanisms
for computing the running-frame, opting instead for rational
(affine) transforms. Because glitch-resilience is the
primary concern of this CLI tool, it uses a "high-water"
approach where the entire ring-buffer is used for safety
padding, and every 10ms the buffer is "topped up" to full
again. Furthermore, because latency is NOT a significant
concern for this CLI tool, the ring-buffer is pre-filled
with silence before the content to be played is added.

Bug: 332322792
Bug: 416750257

Change-Id: I2d442d77129d6e79bf2f8c384a4845aee170fe6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1300168
Reviewed-by: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>