[ftl] Add tracking for possibly worn out blocks.
There is no behavioural change here, just tracking the extra data, a
follow up will set up to export the values.

Adds bitmaps to track the status of blocks that were written during this
mount, indicating that the read count is accurate. So if the block was
recently written and has a relatively low read count then we record the
block as possibly worn.

Bug: 424838073
Change-Id: I447e27d870bb7fd740b13ee97bcf30cf8dd40a12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1300784
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>