[sdio][wifi] Temporary fix to VIM3 Wi-Fi failure
Allow power configuration failure instead of treating it as a
hard error. The result is that power management of SDIO will be
disabled if configuration fails. This means it may be harder to
recognize a failure in situations where we really do expect it
to succeed.

A longer-term fix is needed.

Fixed: 424517899

Change-Id: Ib39a195dd1c84e4aaff075872b40e366c99c5cee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1302334
Commit-Queue: Justin Mattson <jmatt@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>