[archivist] Fix the lazy-node integration test
The test previously passed because the lazy data was being written to a
different Inspector than all the other data. This was a quirk of the
puppet API. I modified the FIDL to have lazy nodes be part of the
InspectWriter protocol, so they can be added to an existing Inspector
rather than started as a new tree.

Bug: 409822059
Fixed: 363007034
Change-Id: I8f3e0d06f257664795fe60e57de9dcc1906b2e1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1250125
Fuchsia-Auto-Submit: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Clayton McCray <claytonmccray@google.com>