[starnix] Drop PartialOrd for SchedulerPolicyKind
Since recently-authored https://fxrev.dev/1247910 it had been in
violation of "The methods of this trait must be consistent with each
other and with those of PartialEq" (and the six conditions that follow
it) at https://doc.rust-lang.org/std/cmp/trait.PartialOrd.html.

Since the functionality is currently only used from Binder, and since
it differs from other methods of comparing schedulings that we might
want to use in other contexts (or even in the context of Binder), we
rename it "is_less_than_for_binder" and punt the question of what about
it to change to https://fxbug.dev/425726327.

Bug: 397371605
Bug: 425726327
Change-Id: Ie490a7a2b78b0cc6d80adba6a8a89c8085594cb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1301645
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Fuchsia-Auto-Submit: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>