[ffx][usb] Flow control
This introduces an advisory flow control packet which sets a flow control state
to either on or off. If we receive a packet turning the state on for a given
connection, we stop reading from the socket for that connection and thus stop
sending data until flow control is turned off again.

Fixed: 422193654
Change-Id: Ib8c52fe0770246d5cd443686784aa90880aab997
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1293650
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>