[component_manager] Fix inactive type checks
Routers now support and apply type checks for porcelain types. However,
there was some logic that was disabling these checks for all types
except for Protocol and Config. I believe this logic is incorrect
because we now support more porcelain types than that in bedrock.

I also had to make a change to the routing logic. When performing a
lookup in a nested path, we have to set the request type for all but the
last component to Dictionary. Otherwise the type check will detect a
mismatch (unless the capability type was Dictionary to begin with).

Change-Id: Ibd8aa193768c8167419e36966abf6d82872e82aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1303284
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Gary Bressler <geb@google.com>