[kernel][vm] Dump information on reclamation hangs
Attempts to detect and dump additional information when the reclamation
system might be wedged. This involves both adding detection to loops
for scenarios where they might be livelocked and excessively spinning
and backtracing eviction threads to find out if they are blocked and
where.

Bug: 422077273
Change-Id: I0650ce102252fa05a599c5aa4292bfb2fe6ddab0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304844
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>