[fshost] Add short-term workaround to address flake
When the Starnix instance dies, fxfs will notice and force lock the
volume. fshost maintains a connection to the volume so when there's
another attempt to mount the volume, it will try to unmount the volume
first which fails because it was forcibly locked.

To workaround this for now, just log an error if the unmount fails.  The
mount should succeed.

Bug: 422152987
Change-Id: I824ef5321f0f2ac7f2d7c9f0daefde7aebcd1f5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1306244
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>