[fxfs] On memory critical only flush files with dirty pages.
Makes a cheaper check to see if there could be any dirty pages when
memory is critical rather than the full complexity of a flush call.

When memory is critical we'd rather get out of the way faster and only
do the work required to lower memory. For the majority of open files,
they don't have anything to flush. The actual flush call needs to be
careful not to miss anything in a complex set of races, and as a result
just checking if a flush is necessary requires making a sys call to get
the vmo stats. When we're trying to reduce memory, all those races don't
really matter and we don't actually care about flushing metadata since
that doesn't recover any memory.

Also moves cache resize up in the background run order, as low memory
situations may prompt it to close a bunch of files before we iterate all
of them.

Bug: b/421199208

Change-Id: Ica0164ed5969b3c1933a6257ab05121fe7d4c34e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304282
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>