[storage] Document storage_device and block_client
Clarifies the different purposes of the 'storage_device' and
'block_client' libraries:

'block_client' is the low-level client for fuchsia.hardware.block.

'storage_device' is a portable abstraction for interacting with block
devices, which can be used on top of `block_client`.  It also includes
an allocator for buffers out of a shared transfer VMO.

Also makes a few miscellaneous doc improvements.

Bug: 424202905
Change-Id: I7c9a89cb39e10d9152511ee6241ce4469b8da45a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1305724
Reviewed-by: Brandon Castellano <bcastell@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>