[trace] Improving lock usage for writing ATrace events.
This change shortens the spans that are locked when
writing ATrace events to Starnix.

This also changes the enabled flag on the TraceEventQueue
to be an atomic value vs. having to lock a mutex to check for
is_enabled which is done very often.

This change reduced the average duration of the write from
.344 ms to .04 ms.

The max write time was reduced from 659 ms to 19 ms.

Bug: 414642814

Change-Id: I9c1c4ad13a6bc68c65732615a486d2bab924cd05
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304684
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Guocheng Wei <guochengwei@google.com>