[lacewing] Honor exit codes from Mobly tests
Surface the exit code of the subprocess in which Lacewing runs Mobly
tests. Until now, Lacewing has assumed that Mobly will always exit 1 on
failure (which turns out to be true), but this is not guaranteed. There
are cases where infrastructure would like to be able to respond
differently to different kinds of test errors and failures. This change
is the first step in that direction.

Bug: b/425675837
Test: fx test //src/testing/end_to_end/mobly_driver/tests/unit_tests --host --output

Change-Id: I43dcad6c6e4a3ae987b3ade4e1998d90708d12d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1307127
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>