[super_parser] Parse super using Arc<dyn Device>
Parse the super information use `Arc<dyn Device>` instead of
`FakeDevice`. This will be useful in a follow-up CL where we would want
to store the device in the parser struct and hand out copies of it.

It also means that the parser isn't restricted to only reading data from
a FakeDevice.

Bug: 404952286
Change-Id: I3dcc01af7334527e606fe83158bc211093785c06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304105
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>