[super_parser] Separate slot partition metadata
The term "metadata" is overloaded in Super. There is the metadata of
"super" which includes the geometry and some number of metadata of
partitions, extents, partition groups, and block devices.

Refactor the library such that there is now one "SuperMetadata" which
contains the disk geometry and copies of the "metadata_slots" which
refers to metadata of the partitions, extents, partition groups, and
block devices. For each "slot", there is a separate metadata. For
example, A/B devices will have two slots.

This refactor should make it easier to understand the super metadata
layout. It also include slight performant changes where that the
geometry information would not be re-read from the device each time a
metadata (for a different slot) is loaded.

Bug: 404952286
Change-Id: I75270a7165ca63f434d796afc81f8d3cda6351e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304106
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>