[sdk] Add `transitive_features` attribute to fuchsia_cc_binary() and fuchsia_cc_test().
Add a new attribute to the fuchsia_cc_xxx() macros to provide a list of
feature flag names to apply to the C++ binary and all its transitive
dependencies. This injects a rule to apply a transition modifying the build
configuration's features list (i.e. `//command_line_option:features`).

This uses a custom rule that can export the CcInfo and CcSharedLibraryInfo
providers from the real C++ binary target, as they are needed by the
fuchsia_cc() rule to attach extra

This is necessary to disable sanitizer features when building certain
targets. The root problem is that due to b/417952817, sanitizer flags are
not passed properly to Bazel when invoked from GN, and trying to fix this
reveals that some //vendor/* Bazel tests are failing at runtime with
undefined behavior issues.

See tqr/1005170 which uses this new attribute to disable asan and ubsan
for the problematic target, then https://fxrev.dev/1269985 which fixes the
flag-passing bug.

Tested also with `fx bazel build @prebuilt_clang//:test_suite`.

Bug: 417952817
Bug: 425323223
Change-Id: Iea555418d944283509b7bcbe4ac655673d1195a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1301585
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>