[sdmmc][core] Ignore initial transitions to OFF and BOOT
Power Framework will immediately move the SDMMC PE to OFF and then to
BOOT after the driver starts. The device should be kept on and the
transitions ignored, otherwise the entire eMMC probe sequence will have
to be repeated on Moonflower.

This case can be removed after b/424264756 is implemented and we can get
a lease at the time the PE is created.

Bug: 422196233
Bug: 424264756
Test: sdmmc-block-test
Change-Id: I66ec0c553f4cabad6aef51fe43d0f4831a45a8e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1305784
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>
Fuchsia-Auto-Submit: Braden Kell <bradenkell@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>