[power][testing] Match on suffix of driver URL in mali test
This updates the msd-arm-mali power integration test to identify the
aml-gpu.cm driver by a suffix of its URL instead of the full URL
including scheme. This is to prepare for moving storage for this driver
from bootfs to a base package which changes the scheme.

Bug: 425682108
Change-Id: I8a6545bb1f2afc02a9eb9fa439b501d3e00c4fb7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1307126
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: James Robinson <jamesr@google.com>