[kernel][vm] Only MarkAccessed after ownership check
When we call MarkAccessed after a failed reclamation attempt, we should
be sure that the page is owned by the VmCowPages. Otherwise we don't
know what state the page is in to safely be able to mark it accessed.
Some paths in ReclaimPageForCompression had this order of operations
flipped.

Bug: 423900604
Run-All-Tests: true
Change-Id: I9e852ebb91af0294a6d7e9ff87a239eb3712cc87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1308010
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>